Initial Setup, settings, and assets
For this first push, I got the project prepped with familiar settings from the last version of Ruins of Vha for the editor, build settings, etc. I also got the input system package and text mesh pro properly set up, and started the file structure. I imported the first round of sprites and the font I intend to use, along with an event system scriptable object setup that I've used previously, and added a credits script, so that I don't lose track of what I've added.