Set up basic movement, character controller, and tile map
Added the PlayerController and PlayerMovement scripts, which take messages from the input system and control the movement of the player respectively. Also re-set the anchor points of the player, shadow, and skeleton sprites to center-bottom to make animating, collisions, and dynamic layer sorting easier. Added a grid and set up the first tile palette.