Added dynamic sprite layering, damaging hazards, invulnerability, and knockback
Added PlayerLayer and ObjectLayer scripts to adjust the sprite renderer's order in layer dynamically based on y position. Also added foreground and background layers to prevent moving sprites from phasing behind the floor or walls visually. Added the ContactHazard and PlayerHealth_Movement scripts for setting up a basic spikes object, and set up methods for triggering a damage animation, an invulnerability window, and a couple types of knockback. The game doesn't track actual damage yet, but the behaviors and animations are working consistently so far.