Started working on enemy behavior
Created a basic enemy game object group, and started working on the EnemyBehavior script. So far I've got the bones of a state machine for what "mode" the enemy is in (patrol, pursuit, returning, etc), and a movement behavior that causes it to pursue it's target position at whatever movement speed is correct for the state. There aren't triggers for changing states set up other than a start function to get it into patrol from idle, but that'll be the next steps.