Added player detection and chasing to enemy behavior
Short update today to get back into the swing of things. Added player detection to the skeleton enemy's alert collider, and added methods to transition into the alert and chasing states for the skeleton. Haven't added the alert animation yet, and the chase speed is currently very fast, but will add the other parts of the skeleton chase behavior next session.