Added enemy damage, alert and gloat timers, and dismissal
The enemy will now damage and knock-back the player when it contacts them, and there is now a timer for the alert state (a grace period to let the player get a head start to run away). The enemy transitions to a gloating state after damaging the player, which also gives the player a grace period to flee. Lastly, the chase limiter script prevents the enemy from chasing the player further than a fixed distance away from it's patrol zone. Also added an additional alert area that can transition out of the dismissal state to prevent the player from cheesing the state change threshold for dismissal. Might add a similar detection script for the patrol zone itself, but next steps will be to add animation and rotating the detection areas for the enemy.